# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReMarket API
  version: 1.0.0
  description: |
    REST contract for the ReMarket marketplace. Monetary values are decimal
    strings in VND and timestamps are ISO-8601 UTC strings. Mutating order and
    product operations use optimistic versions. Checkout requires an
    Idempotency-Key header and replays the original result for the same payload.
servers:
  - url: /api/v1
...
paths:
...
  /admin/email-verifications:
    get:
      tags: [Admin]
      operationId: adminEmailVerifications
      security: [{ bearerAuth: [] }]
      description: ACTIVE ADMIN only. Pending requests by default, including users who became locked after submitting.
      parameters: [{ name: status, in: query, schema: { type: string, enum: [PENDING, APPROVED, ALL], default: PENDING } }, { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }]
      responses:
        '200': { $ref: '#/components/responses/EmailVerificationListSuccess' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
  /admin/email-verifications/{id}/approve:
    post:
      tags: [Admin]
      operationId: adminApproveEmailVerification
      security: [{ bearerAuth: [] }]
      description: ACTIVE ADMIN approves an existing request. Idempotent under concurrency. Updates only emailVerifiedAt; does not unlock accounts or grant roles. Approval, audit, notification and realtime outbox commit atomically.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, additionalProperties: false, example: {} } } } }
      responses:
        '200': { $ref: '#/components/responses/EmailVerificationSuccess' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/Error' }
...
  /conversations/{conversationId}/read:
    post: { tags: [Chat], operationId: markConversationRead, security: [{ bearerAuth: [] }], parameters: [{ name: conversationId, in: path, required: true, schema: { type: string, format: uuid } }], requestBody: { $ref: '#/components/requestBodies/MarkReadBody' }, responses: { '200': { $ref: '#/components/responses/MarkReadSuccess' } } }
...
  /reports/mine:
    get: { tags: [Trust], operationId: listOwnReports, security: [{ bearerAuth: [] }], parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/PageSize' }], responses: { '200': { $ref: '#/components/responses/ReportListSuccess' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  requestBodies:
...
    MarkReadBody:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/MarkReadInput' } } }
...
  responses:
    EmailVerificationSuccess:
      description: Own request or approved request; nullable only for accounts without a request.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/SuccessEnvelope' }
              - { type: object, properties: { data: { allOf: [{ $ref: '#/components/schemas/EmailVerificationRequest' }], nullable: true } } }
    EmailVerificationListSuccess:
      description: Paginated requests available only to ACTIVE admins.
      content:
        application/json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/SuccessEnvelope' }
              - { type: object, properties: { data: { type: object, required: [items, meta], properties: { items: { type: array, items: { $ref: '#/components/schemas/EmailVerificationRequest' } }, meta: { $ref: '#/components/schemas/PageMeta' } } } } }
...
    MarkReadSuccess:
      description: Conversation read state
      content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }, { type: object, properties: { data: { type: object, required: [unread, updated, read_at], additionalProperties: false, properties: { unread: { type: integer, minimum: 0 }, updated: { type: integer, minimum: 0 }, read_at: { $ref: '#/components/schemas/IsoDateTime' } } } } }] } } }
...
    ReportListSuccess:
      description: Paginated reports
      content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/SuccessEnvelope' }, { type: object, properties: { data: { $ref: '#/components/schemas/ReportListData' } } }] } } }
...
    Error:
      description: Business or validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    EmailVerificationRequest:
      type: object
      required: [id, user, status, requested_at, approved_at]
      properties:
        id: { type: string, format: uuid }
        user:
          type: object
          required: [id, full_name, email, status]
          properties:
            id: { type: string, format: uuid }
            full_name: { type: string }
            email: { type: string, format: email }
            status: { type: string, enum: [ACTIVE, LOCKED] }
        status: { type: string, enum: [PENDING, APPROVED] }
        requested_at: { type: string, format: date-time }
        approved_at: { type: string, format: date-time, nullable: true }
...
    MarkReadInput:
      type: object
      required: [last_message_id]
      additionalProperties: false
      properties: { last_message_id: { type: string, format: uuid } }
...
    IsoDateTime: { type: string, format: date-time }
...
    ReportReason: { type: string, enum: [COUNTERFEIT, PROHIBITED, FRAUD, SPAM, HARASSMENT, INAPPROPRIATE, OTHER] }
    ReportStatus: { type: string, enum: [PENDING, RESOLVED, REJECTED] }
...
    PageMeta:
      type: object
      required: [page, page_size, total, total_pages]
      additionalProperties: false
      properties:
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        total_pages: { type: integer, minimum: 1 }
...
    ReportRecord:
      type: object
      required: [id, target_type, target_label, target_id, reason, description, status, resolution_note, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        target_type: { type: string, enum: [product, user] }
        target_label: { type: string }
        target_id: { type: string, format: uuid }
        reason: { $ref: '#/components/schemas/ReportReason' }
        description: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/ReportStatus' }
        resolution_note: { type: string, nullable: true }
        created_at: { $ref: '#/components/schemas/IsoDateTime' }
        updated_at: { $ref: '#/components/schemas/IsoDateTime' }
...
    ReportListData:
      type: object
      required: [items, meta]
      additionalProperties: false
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/ReportRecord' } }
        meta: { $ref: '#/components/schemas/PageMeta' }
...
    Meta:
      type: object
      required: [request_id]
      properties:
        request_id: { type: string, format: uuid }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
        total_pages: { type: integer, minimum: 1 }
    SuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        meta: { $ref: '#/components/schemas/Meta' }
    ApiErrorCode:
      type: string
      description: Stable machine-readable error code shared by the API and web client.
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - FORBIDDEN
        - EMAIL_NOT_VERIFIED
        - ACCOUNT_LOCKED
        - NOT_FOUND
        - PRODUCT_NOT_AVAILABLE
        - PRICE_CHANGED
        - INVALID_ORDER_TRANSITION
        - ORDER_EXPIRED
        - VERSION_CONFLICT
        - REVIEW_NOT_ALLOWED
        - IDEMPOTENCY_CONFLICT
        - RETRY_LATER
        - RATE_LIMITED
        - SESSION_EXPIRED
        - INTERNAL
      example: VERSION_CONFLICT
    ErrorEnvelope:
      type: object
      required: [success, error, meta]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message]
          properties:
            code: { $ref: '#/components/schemas/ApiErrorCode' }
            message: { type: string, example: Dữ liệu đã thay đổi, vui lòng tải lại. }
            details: { type: object, additionalProperties: true }
        meta: { $ref: '#/components/schemas/Meta' }